feat: 쓰레기통 기능 구현
WasteBin 스크립트 작성
- OnTriggerEnter : Collider 로 Burger 종류이면 해당 오브젝트 Destroy
