+++ Assets/Scenes/yourzinc/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d1becfb2c310054f99a58907535f5f9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
