+++ Assets/Scenes/yourzinc/Scripts/PlateController.cs.meta
