+++ Assets/Scenes/yourzinc/Scripts/WasteBin.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 09fd2a37c88da80478dc33d082bbc461
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
