+++ Assets/Scenes/yourzinc/yourzinc0524.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 05716c1f670a85349a797d053d60c157
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
